refactor(docker): split depends_on logic by concern
Break the monolithic container_depends_on.go into focused files:

- compose.go: Compose label/condition constants, label parsing, container
  discovery.
- dependency_graph.go: the resolved dependency tree and cycle detection,
  with no Docker dependencies.
- dependency_condition.go: waiting for and evaluating depends_on conditions.
- container_depends_on.go: orchestration that builds the tree and starts it
  in dependency order.

Rename treeDep to dependency, trim comments to explain intent only, and
rename the test files to match their source files. No behaviour change.
